requirements.txt
python-dotenv
flask
flask-cors
gunicorn
